\name{create.EquityIndex}
\alias{create.EquityIndex}
\title{ Create an EquityIndex object }
\description{
}
\usage{
create(EquityIndex, ticker = "^GSPC", hint = NA, src = "yahoo")
}
\arguments{
  \item{EquityIndex}{ The type }
  \item{ticker}{ The ticker representing the index }
  \item{hint}{ }
  \item{src}{ }
}
\details{
}
\value{
An EquityIndex object
}
\author{
Brian Lee Yung Rowe
}

\examples{
  index <- create(EquityIndex)
}
