\name{create.TawnyPortfolio}
\alias{create.TawnyPortfolio}
\alias{create.TawnyPortfolio.ret}
\alias{create.TawnyPortfolio.sym}
\title{ Create a TawnyPortfolio object }
\description{
}
\usage{
create(TawnyPortfolio, returns, window = 90)
create(TawnyPortfolio, symbols, window = 90, obs = 150)
}
\arguments{
  \item{returns}{ }
  \item{symbols}{ }
  \item{window}{ }
  \item{obs}{ }
}
\details{
}
\value{
}
\author{
Brian Lee Yung Rowe
}

\examples{
  port <- create(TawnyPortfolio, c('BAC','JPM','C','GS','MS'))
}
